# Improve performances on queries