--- a/packages/cozy-client/src/CozyClient.js
+++ b/packages/cozy-client/src/CozyClient.js
@@ -1126,6 +1126,73 @@ client.query(Q('io.cozy.bills'))`)
   }
 
   /**
+   * Save the document or array of documents into the persisted storage (if any)
+   *
+   * @private
+   * @param {CozyClientDocument | Array<CozyClientDocument>} data - Document or array of documents to be saved
+   * @returns {Promise<void>}
+   */
+  async persistVirtualDocuments(definition, data) {
+    const enforceList = ['io.cozy.files.shortcuts']
+
+    const enforce = enforceList.includes(definition.doctype)
+
+    if (definition.doctype === 'io.cozy.apps_registry') {
+      // io.cozy.apps_registry has a dedicated `maintenance` endpoint on cozy-stack that
+      // returns data different than the one stored in database
+      // As we want to have transparent queries, whether it uses the stack API or Pouch,
+      // we store the full response into a single doc, with a `maintenance` _id
+      // and a special `cozyPouchData` attribute, to highlight this special case
+      return await this.persistVirtualDocument(
+        {
+          _type: 'io.cozy.apps_registry',
+          _id: definition.id,
+          // @ts-ignore
+          cozyPouchData: data
+        },
+        enforce
+      )
+    }
+
+    if (!Array.isArray(data)) {
+      await this.persistVirtualDocument(data, enforce)
+    } else {
+      const documentsToPersist = data.filter(document => {
+        if (!document || document.cozyLocalOnly) {
+          return false
+        }
+
+        if ((!document.meta?.rev && !document._rev) || enforce) {
+          return true
+        }
+
+        return false
+      })
+      for (const document of documentsToPersist) {
+        await this.persistVirtualDocument(document, enforce)
+      }
+    }
+  }
+
+  /**
+   * Save the document or array of documents into the persisted storage (if any)
+   *
+   * @private
+   * @param {CozyClientDocument} document - Document to be saved
+   * @param {boolean} enforce - When true, save the document even if `meta.rev` or `_rev` exist
+   * @returns {Promise<void>}
+   */
+  async persistVirtualDocument(document, enforce) {
+    if (!document || document.cozyLocalOnly) {
+      return
+    }
+
+    if ((!document.meta?.rev && !document._rev) || enforce) {
+      await this.chain.persistCozyData(document)
+    }
+  }
+
+  /**
    * Fetch relationships for a response (can be several docs).
    * Fills the `relationships` attribute of each documents.
    *