--- a/packages/cozy-pouch-link/src/jsonapi.js
+++ b/packages/cozy-pouch-link/src/jsonapi.js
@@ -1,35 +1,56 @@
-import { generateWebLink } from 'cozy-client'
+import CozyClient, { generateWebLink } from 'cozy-client'
 
-export const normalizeDoc = (doc, doctype, client) => {
-  const id = doc._id || doc.id
-
-  const { relationships, referenced_by } = doc
-
-  // PouchDB sends back .rev attribute but we do not want to
-  // keep it on the server. It is potentially higher than the
-  // _rev.
-  const _rev = doc.rev || doc._rev
-  const normalizedDoc = {
-    ...doc,
-    id,
-    _id: id,
-    _rev,
-    _type: doctype,
-    relationships: {
-      ...relationships,
-      referenced_by
+/**
+ * Normalize several PouchDB document
+ *
+ * @param {CozyClient} client - The CozyClient instance
+ * @param {string} doctype - The document's doctype
+ * @param {Array<import('./CozyPouchLink').CozyClientDocument>} docs - The documents to normalize
+ */
+export const normalizeDocs = (client, doctype, docs) => {
+  for (let i = docs.length; i >= 0; i--) {
+    const doc = docs[i]
+    if (!doc) {
+      docs.splice(i, 1)
+      continue
     }
+    normalizeDoc(client, doctype, doc)
   }
-  if (normalizedDoc.rev) {
-    delete normalizedDoc.rev
-  }
-
-  normalizeAppsLinks(normalizedDoc, doctype, client)
-
-  return normalizedDoc
 }
 
-const normalizeAppsLinks = (docRef, doctype, client) => {
+/**
+ * Normalize a PouchDB document
+ * Note we directly modify the objet rather than creating a new one, as it is
+ * much more performant. See commit description for details.
+ *
+ * @param {CozyClient} client - The CozyClient instance
+ * @param {string} doctype - The document's doctype
+ * @param {import('./CozyPouchLink').CozyClientDocument} doc - The document to normalize
+ */
+export const normalizeDoc = (client, doctype, doc) => {
+  const id = doc._id || doc.id
+
+  doc.id = id
+  doc._id = id
+  doc._rev = doc._rev || doc.rev
+  doc._type = doctype
+  if (doc.relationships) {
+    doc.relationships.referenced_by = doc.referenced_by
+  } else {
+    doc.relationships = {
+      referenced_by: doc.referenced_by
+    }
+  }
+  if (doc.rev) {
+    doc.rev = undefined
+  }
+
+  if (doctype === 'io.cozy.apps') {
+    normalizeAppsLinks(client, doctype, doc)
+  }
+}
+
+const normalizeAppsLinks = (client, doctype, docRef) => {
   if (doctype !== 'io.cozy.apps') {
     return
   }
@@ -50,8 +71,6 @@ const normalizeAppsLinks = (docRef, doctype, client) => {
   }
 }
 
-const filterDeletedDocumentsFromRows = doc => !!doc
-
 export const fromPouchResult = ({ res, withRows, doctype, client }) => {
   // Sometimes, queries are transformed by Collections and they call a dedicated
   // cozy-stack route. When this is the case, we want to be able to replicate the same
@@ -68,22 +87,23 @@ export const fromPouchResult = ({ res, withRows, doctype, client }) => {
   }
 
   if (withRows) {
-    const docs = res.rows
-      ? res.rows.map(row => row.doc).filter(filterDeletedDocumentsFromRows)
-      : res.docs
+    const docs = res.rows ? res.rows.map(row => row.doc) : res.docs
     const offset = res.offset || 0
+    normalizeDocs(client, doctype, docs)
 
-    return {
-      data: docs.map(doc => normalizeDoc(doc, doctype, client)),
+    const result = {
+      data: docs,
       meta: { count: docs.length },
       skip: offset,
       next: offset + docs.length < res.total_rows || docs.length >= res.limit
     }
+    return result
   } else {
+    Array.isArray(res)
+      ? normalizeDocs(client, doctype, res)
+      : normalizeDoc(client, doctype, res)
     return {
-      data: Array.isArray(res)
-        ? res.map(doc => normalizeDoc(doc, doctype, client))
-        : normalizeDoc(res, doctype, client)
+      data: res
     }
   }
 }