--- a/packages/cozy-pouch-link/types/jsonapi.d.ts
+++ b/packages/cozy-pouch-link/types/jsonapi.d.ts
@@ -1,4 +1,5 @@
-export function normalizeDoc(doc: any, doctype: any, client: any): any;
+export function normalizeDocs(client: CozyClient, doctype: string, docs: Array<import('./CozyPouchLink').CozyClientDocument>): void;
+export function normalizeDoc(client: CozyClient, doctype: string, doc: any): void;
 export function fromPouchResult({ res, withRows, doctype, client }: {
     res: any;
     withRows: any;
@@ -6,14 +7,12 @@ export function fromPouchResult({ res, withRows, doctype, client }: {
     client: any;
 }): {
     data: any;
-    meta?: undefined;
-    skip?: undefined;
-    next?: undefined;
-} | {
-    data: any;
     meta: {
         count: any;
     };
     skip: any;
     next: boolean;
+} | {
+    data: any;
 };
+import CozyClient from "cozy-client";