… docs/api/cozy-client/classes/CozyClient.md
… docs/api/cozy-pouch-link/classes/PouchLink.md