This includes improvements to improve queries performances, after we noticed a high cozy-client overhead on large queries.
The details and performances improvements can be found in dedicated commits. 

For an overall performance insight, we measured a big query retrieving 25K documents, on a Redmi Note 10 mobile.
The query is the following: `await client.query(Q('io.cozy.files').limitBy(null))`

The results are : 

Before: 
- 2328ms to read pouch data 
**- Total: 5303ms**
- **43%** of total time is spent on database side

After:
- 2749ms to read pouch data
**- Total: 3156ms**
- **87%** of total time is spent on database side

Thus, before, 57% of the total query time was spent *outside* of the database. Now it is only 13% :rocket: